refactor(oracles): split baseline row models
Why:
- keep baseline row schema declarations separate from parser orchestration before the parser module grows further

What:
- move CoinTracking baseline Pydantic row models into row_models.py
- keep read and parse orchestration in rows.py

Checks:
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pytest -q --no-cov tests/oracles/cointracking/test_baseline_rows.py tests/oracles/workflows/test_baseline_service.py tests/oracles/cointracking/test_baseline.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run ruff check tools/oracles/cointracking/baseline/rows.py tools/oracles/cointracking/baseline/row_models.py
